# [router] improve router logs and request id header