## Motivation

The router was generating verbose logs that made it difficult to debug issues in production environments. Additionally, there was no consistent way to track requests across the distributed system, making it challenging to correlate logs when debugging customer issues. This PR addresses these problems by:

1. Implementing request ID tracking for distributed tracing
2. Reducing log verbosity while maintaining important operational information
3. Making request ID headers configurable to support different platforms and standards

## Modifications

### 1. Request ID Tracking with Middleware
- Implemented `RequestIdMiddleware` using idiomatic Actix-web patterns
- Automatically extracts request IDs from configurable HTTP headers (defaults: `x-request-id`, `x-correlation-id`, `x-trace-id`, `request-id`)
- Generates UUID v4 when no request ID is found in headers
- Stores request ID in request extensions for consistent access across all handlers
- Made request ID headers configurable via:
  - Python API: `Router(..., request_id_headers=["x-custom-id"])`
  - CLI: `--request-id-headers x-custom-id x-trace-id`

### 2. Improved Logging Throughout the Codebase
- Added request ID to all HTTP request logs for end-to-end traceability
- Converted verbose multi-line logs to concise, structured formats
- Used tracing's structured logging with named fields for better parsing:
  ```rust
  info!(
      request_id = %request_id,
      route = %route,
      worker_url = %worker_url,
      "Routing request"
  );
  ```

### 3. Reduced Log Noise
- Moved routine operational logs from `info` to `debug` level:
  - Health check logs during startup
  - Load monitoring updates
  - Cache eviction logs
- Kept important state changes at `info` level:
  - Worker additions/removals
  - Request routing decisions
  - Errors and retries

### 4. Performance Analysis
Comprehensive benchmarking shows the middleware has negligible impact:
- **Synthetic benchmark**: 220ns overhead per request (40% relative to bare minimum)
- **Realistic benchmark**: 12µs overhead per request (**0.55% relative to typical 1ms processing**)
- **Conclusion**: Production-ready with <1% performance impact in real scenarios

### Files Modified
- `src/server.rs` - Added RequestIdMiddleware implementation
- `src/lib.rs` - Added request_id_headers configuration field
- `src/config/types.rs` - Updated RouterConfig struct
- `src/routers/router.rs` - Added request ID logging to routing logic
- `src/routers/pd_router.rs` - Added request ID to PD routing logs
- `py_src/sglang_router/router.py` - Added Python API support
- `py_src/sglang_router/launch_router.py` - Added CLI argument

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [x] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.

## Benchmark Results

### Performance Impact
The request ID middleware adds minimal overhead:
- **With header present**: +12µs per request (0.55% overhead)
- **Without header (UUID generation)**: Similar performance due to measurement variance
- **Memory impact**: Negligible (one string per request)

### Test
1. **Synthetic test**: 10,000 requests with no processing to measure pure middleware overhead
2. **Realistic test**: 1,000 requests with 1ms processing time to simulate actual router workload

The realistic benchmark shows that in production conditions, the middleware overhead is negligible (<1%), making it suitable for high-throughput environments.